Fixes #2041. Check if the folder 'docs' exists before removing in build.ps1.

+++ docfx/build.ps1
@@ -2,7 +2,10 @@
 
 dotnet build --configuration Release ../Terminal.sln
 
-rm ../docs -Recurse -Force
+$Folder = '..\docs'
+if(Test-Path -Path $Folder){
+	rm ../docs -Recurse -Force
+}
 
 $env:DOCFX_SOURCE_BRANCH_NAME="main"
